src/google_iot_core_gateway/requirements.txt
colorlog==5.0.1
paho-mqtt==1.5.0
pyjwt==2.1.0
cryptography==3.3.1
google-cloud-pubsub==2.6.0
google-cloud-iot==2.1.0
google-api-python-client==2.12.0
# google-auth-httplib2==0.1.0
# google-auth==1.31.0
# grpc-google-iam-v1==0.12.3
